Initial Upload of project code
This is to be put into the official  code under the script folder. To change the model, simply change it in the bottom of the code with test_classifier method. I also  commented out the csv dump code in the bottom of the code. Uncomment to dump database to csv